requirements.txt
openai==1.3.7
streamlit==1.28.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
